requirements.txt
asgiref==3.11.0
attrs==25.4.0
dj-database-url==1.0.0
Django==5.2.8
django-ckeditor==6.5.0
django-cors-headers==3.11.0
django-crontab==0.7.1
django-environ==0.9.0
django-js-asset==2.2.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
drf-spectacular-sidecar==2025.10.1
Faker==38.2.0
gunicorn==20.1.0
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
legacy-cgi==2.6.4
pillow==12.0.0
psycopg2-binary==2.9.9
PyJWT==2.10.1
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.29.0
setuptools==80.9.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
whitenoise==6.2.0
